+++ homekit/discover_ble.py
@@ -15,135 +15,16 @@
 # limitations under the License.
 #
 
-import threading
-import time
 import logging
 from argparse import ArgumentParser
-import os
-from homekit.model import Categories
 
-from staging import gatt
-
-
-class Killer(threading.Thread):
-
-    def __init__(self, manager, timeout):
-        threading.Thread.__init__(self)
-        self.timeout = timeout
-        self.manager = manager
-
-    def run(self):
-        time.sleep(self.timeout)
-        self.manager.stop()
-
-
-def parse_manufacturer_specific(input_data):
-    logging.debug('manufacturer specific data: %s', input_data.hex())
-
-    # the type must be 0x06 as defined on page 124 table 6-29
-    ty = input_data[0]
-    input_data = input_data[1:]
-    if ty == 0x06:
-        ty = 'HomeKit'
-
-        ail = input_data[0]
-        logging.debug('advertising interval %s', '{0:02x}'.format(ail))
-        length = ail & 0b00011111
-        if length != 13:
-            logging.debug('error with length of manufacturer data')
-        input_data = input_data[1:]
-
-        sf = input_data[0]
-        if sf == 0:
-            sf = 'paired'
-        elif sf == 1:
-            sf = 'unpaired'
-        else:
-            sf = 'error'
-        input_data = input_data[1:]
-
-        device_id = (':'.join(input_data[:6].hex()[0 + i:2 + i] for i in range(0, 12, 2))).upper()
-        input_data = input_data[6:]
-
-        acid = int.from_bytes(input_data[:2], byteorder='little')
-        input_data = input_data[2:]
-
-        gsn = int.from_bytes(input_data[:2], byteorder='little')
-        input_data = input_data[2:]
-
-        cn = input_data[0]
-        input_data = input_data[1:]
-
-        cv = input_data[0]
-        input_data = input_data[1:]
-        if len(input_data) > 0:
-            logging.debug('remaining data: %s', input_data.hex())
-        return {'manufacturer': 'apple', 'type': ty, 'sf': sf, 'deviceId': device_id, 'acid': acid,
-                'gsn': gsn, 'cn': cn, 'cv': cv, 'category': Categories[int(acid)]}
-
-    return {'manufacturer': 'apple', 'type': ty}
-
-
-class Device(gatt.Device):
-
-    homekit_discovery_data = None
-
-    def __init__(self, *args, **kwargs):
-        gatt.Device.__init__(self, *args, **kwargs, managed=False)
-
-        self.name = self._properties.Get('org.bluez.Device1', 'Alias')
-        self.homekit_discovery_data = self.get_homekit_discovery_data()
-
-    def get_homekit_discovery_data(self):
-        import dbus
-
-        try:
-            mfr_data = self._properties.Get('org.bluez.Device1', 'ManufacturerData')
-        except dbus.exceptions.DBusException as e:
-            if e.get_dbus_name() == 'org.freedesktop.DBus.Error.InvalidArgs':
-                return {}
-            raise
-
-        mfr_data = dict((int(k), bytes(bytearray(v))) for (k, v) in mfr_data.items())
-
-        if 76 not in mfr_data:
-            return {}
-
-        parsed = parse_manufacturer_specific(mfr_data[76])
-
-        if parsed['type'] != 'HomeKit':
-            return
-
-        return parsed
-
-    #def advertised(self):
-    #    print(self.get_homekit_discovery_data())
-
-    def __repr__(self):
-        return 'BleDevice[mac_address="{}", name="{}"]'.format(self.mac_address, self.name)
-
-
-class DeviceManager(gatt.DeviceManager):
-
-    discover_callback = None
-
-    def make_device(self, mac_address):
-        device = Device(mac_address=mac_address, manager=self)
-        if not device.homekit_discovery_data:
-            return
-        self._manage_device(device)
-        if self.discover_callback:
-            self.discover_callback(device)
-
-    def start_discovery(self, callback=None):
-        self.discover_callback = callback
-        return gatt.DeviceManager.start_discovery(self)
+from homekit.controller.ble_impl.discovery import DiscoveryDeviceManager
 
 
 def discover(adapter, timeout=10):
-    manager = DeviceManager(adapter)
+    manager = DiscoveryDeviceManager(adapter)
     manager.start_discovery()
-    Killer(manager, timeout).start()
+    manager.set_timeout(timeout * 1000)
     manager.run()
 
     return manager._devices.values()
