--- staging/gatt/__init__.py
@@ -1,1 +1,0 @@
-from .gatt import DeviceManager, Device, Service, Characteristic
