--- staging/gatt/errors.py
@@ -1,30 +1,0 @@
-class AccessDenied(Exception):
-		pass
-
-
-class Failed(Exception):
-        pass
-
-
-class InProgress(Exception):
-        pass
-
-
-class InvalidValueLength(Exception):
-		pass
-
-
-class NotAuthorized(Exception):
-        pass
-
-
-class NotReady(Exception):
-        pass
-
-
-class NotPermitted(Exception):
-        pass
-
-
-class NotSupported(Exception):
-		pass
