--- staging/gatt/gatt_stubs.py
@@ -1,14 +1,0 @@
-class DeviceManager:
-    pass
-
-
-class Device:
-    pass
-
-
-class Service:
-    pass
-
-
-class Characteristic:
-    pass
